fix(python): upgrade PyO3 to 0.29
Resolve the PyO3 iterator and closure-safety advisories while preserving native and Pyodide wheel support. Align stub generation with the private extension module and validate public re-exports.
