+++ Cargo.toml
@@ -48,7 +48,7 @@ smallvec = "1.13"
 rustc-hash = "1.1"
 tempfile = "3.10"
 web-time = "1.1.0"
-pyo3 = "0.27.1"
+pyo3 = "0.29.0"
 wasm-bindgen = "0.2.95"
 wasm-bindgen-futures = "0.4.45"
 js-sys = "0.3.72"
