+++ bindings/python/formualizer/_mypy_smoke.py
@@ -4,10 +4,36 @@ import datetime
 from collections.abc import Mapping
 from typing import Any
 
-from . import SheetPortSession, Workbook
+from . import (
+    ExcelEvaluationError,
+    FormualizerHostError,
+    ParserError,
+    SheetPortConstraintError,
+    SheetPortError,
+    SheetPortManifestError,
+    SheetPortSession,
+    SheetPortWorkbookError,
+    TokenizerError,
+    Workbook,
+)
+from .formualizer_py import Workbook as NativeWorkbook
+
+_EXCEPTION_TYPES: tuple[type[Exception], ...] = (
+    TokenizerError,
+    ParserError,
+    FormualizerHostError,
+    ExcelEvaluationError,
+    SheetPortError,
+    SheetPortManifestError,
+    SheetPortConstraintError,
+    SheetPortWorkbookError,
+)
 
 
 def _mypy_api_smoke(wb: Workbook, session: SheetPortSession) -> Mapping[str, Any]:
+    # The public package is a typing re-export facade, not a second nominal API.
+    native_wb: NativeWorkbook = wb
+
     # Workbook changelog metadata setters
     wb.set_actor_id(None)
     wb.set_correlation_id("corr")
@@ -27,4 +53,4 @@ def _mypy_api_smoke(wb: Workbook, session: SheetPortSession) -> Mapping[str, Any
     )
 
     # Return a value so mypy checks mapping types.
-    return {"out1": out1, "out2": out2}
+    return {"out1": out1, "out2": out2, "native_wb": native_wb}
