+++ bindings/python/pyproject.toml
@@ -48,8 +48,8 @@ dev = [
 # Keep Rust crate/module named `formualizer_py`, but install it under the
 # public Python package `formualizer` to avoid top-level module conflicts.
 module-name = "formualizer.formualizer_py"
-# Mark as mixed Rust/Python layout so stub generation can place `__init__.pyi`
-# into the correct package directory.
+# Mark as mixed Rust/Python layout. The stub generator promotes native-module
+# declarations to the public package and writes a companion native-module stub.
 python-source = "."
 features = ["pyo3/extension-module", "pyo3/abi3-py310"]
 
@@ -59,6 +59,7 @@ line-length = 88
 exclude = [
     # Generated by `cargo run --bin stub_gen`
     "formualizer/__init__.pyi",
+    "formualizer/formualizer_py.pyi",
 ]
 
 [tool.ruff.lint]
