+++ bindings/python/src/bin/stub_gen.rs
@@ -1,24 +1,102 @@
+#[cfg(not(target_os = "emscripten"))]
 use pyo3_stub_gen::Result;
+#[cfg(not(target_os = "emscripten"))]
 use std::fs;
+#[cfg(not(target_os = "emscripten"))]
 use std::path::PathBuf;
 
+#[cfg(target_os = "emscripten")]
+fn main() {}
+
+#[cfg(not(target_os = "emscripten"))]
 fn main() -> Result<()> {
     // `stub_info` is defined in `src/lib.rs` by `define_stub_info_gatherer!`.
     let stub = formualizer_py::stub_info()?;
     stub.generate()?;
 
-    // Post-process the generated package stub to reflect the public Python API aliases
-    // defined in `formualizer/__init__.py`.
-    let stub_path: PathBuf = [env!("CARGO_MANIFEST_DIR"), "formualizer", "__init__.pyi"]
+    // pyo3-stub-gen 0.23 follows maturin's private extension module path and
+    // writes `formualizer/formualizer_py/__init__.pyi`. The public Python package
+    // re-exports that module, so publish the declarations as a companion
+    // `formualizer_py.pyi`, emit a public re-export facade at the package root,
+    // and remove the intermediate package-shaped directory.
+    let generated_stub_path: PathBuf = [
+        env!("CARGO_MANIFEST_DIR"),
+        "formualizer",
+        "formualizer_py",
+        "__init__.pyi",
+    ]
+    .iter()
+    .collect();
+    let generated_stub_dir = generated_stub_path
+        .parent()
+        .expect("generated stub path has a parent");
+    let private_stub_path: PathBuf = [
+        env!("CARGO_MANIFEST_DIR"),
+        "formualizer",
+        "formualizer_py.pyi",
+    ]
+    .iter()
+    .collect();
+    let public_stub_path: PathBuf = [env!("CARGO_MANIFEST_DIR"), "formualizer", "__init__.pyi"]
         .iter()
         .collect();
 
-    let mut contents = fs::read_to_string(&stub_path)?;
+    let mut contents = fs::read_to_string(&generated_stub_path)?;
+    fs::remove_dir_all(generated_stub_dir)?;
 
-    // Idempotent append.
-    if !contents.contains("# Backwards compatible Py* aliases") {
-        contents.push_str(
-            r#"
+    for (generated, corrected) in [
+        (
+            "def __getitem__(self, index: builtins.int)",
+            "def __getitem__(self, index: builtins.int, /)",
+        ),
+        (
+            "def __getitem__(self, name: builtins.str)",
+            "def __getitem__(self, name: builtins.str, /)",
+        ),
+    ] {
+        assert!(
+            contents.contains(generated),
+            "generated stub is missing the expected magic-method signature: {generated}"
+        );
+        contents = contents.replacen(generated, corrected, 1);
+    }
+
+    let additional_public_exports = r#"    "ExcelEvaluationError",
+    "FormualizerHostError",
+    "ParserError",
+    "PyFormulaDialect",
+    "PyRefWalker",
+    "PyToken",
+    "PyTokenSubType",
+    "PyTokenType",
+    "PyTokenizer",
+    "PyTokenizerIter",
+    "SheetPortConstraintError",
+    "SheetPortError",
+    "SheetPortManifestError",
+    "SheetPortWorkbookError",
+    "TokenizerError",
+"#;
+    let all_end = "]\n\n@typing.final";
+    assert!(
+        contents.contains(all_end),
+        "generated stub is missing the expected __all__ terminator"
+    );
+    contents = contents.replacen(all_end, &format!("{additional_public_exports}{all_end}"), 1);
+
+    contents.push_str(
+        r#"
+
+# Exceptions created with pyo3::create_exception! are public runtime types but
+# are not currently discovered by pyo3-stub-gen's inventory.
+class TokenizerError(Exception): ...
+class ParserError(Exception): ...
+class FormualizerHostError(Exception): ...
+class ExcelEvaluationError(Exception): ...
+class SheetPortError(Exception): ...
+class SheetPortManifestError(SheetPortError): ...
+class SheetPortConstraintError(SheetPortError): ...
+class SheetPortWorkbookError(SheetPortError): ...
 
 # Backwards compatible Py* aliases
 #
@@ -32,9 +110,35 @@ PyTokenType = TokenType
 PyTokenSubType = TokenSubType
 PyFormulaDialect = FormulaDialect
 "#,
-        );
-        fs::write(&stub_path, contents)?;
-    }
+    );
+    let all_start = contents
+        .find("__all__ = [")
+        .expect("generated stub is missing __all__");
+    let all_close = all_start
+        + contents[all_start..]
+            .find("]\n\n@typing.final")
+            .expect("generated stub is missing the end of __all__");
+    let mut public_all = contents[all_start..=all_close].to_owned();
+    public_all.insert_str(
+        public_all.len() - 1,
+        "    \"ReferenceLike\",\n    \"visitor\",\n",
+    );
+
+    fs::write(&private_stub_path, contents)?;
+    fs::write(
+        &public_stub_path,
+        format!(
+            r#"# This file is automatically generated by pyo3_stub_gen
+# ruff: noqa: E501, F401, F403, F405
+
+from . import visitor as visitor
+from ._types import ReferenceLike as ReferenceLike
+from .formualizer_py import *
+
+{public_all}
+"#
+        ),
+    )?;
 
     Ok(())
 }
