+++ bindings/python/src/enums.rs
@@ -3,11 +3,16 @@ use std::fmt::Display;
 use formualizer::parse::tokenizer::{TokenSubType as CoreTokenSubType, TokenType as CoreTokenType};
 use formualizer::parse::types::FormulaDialect as CoreFormulaDialect;
 use pyo3::prelude::*;
+#[cfg(not(target_os = "emscripten"))]
 use pyo3_stub_gen::derive::{gen_stub_pyclass_enum, gen_stub_pymethods};
 
 /// Python-exposed token type enum
-#[gen_stub_pyclass_enum]
-#[pyclass(name = "TokenType", module = "formualizer")]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pyclass_enum)]
+#[pyclass(
+    name = "TokenType",
+    module = "formualizer.formualizer_py",
+    from_py_object
+)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum PyTokenType {
     Literal,
@@ -28,7 +33,7 @@ impl Display for PyTokenType {
     }
 }
 
-#[gen_stub_pymethods]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pymethods)]
 #[pymethods]
 impl PyTokenType {
     fn __str__(&self) -> &'static str {
@@ -86,8 +91,12 @@ impl From<PyTokenType> for CoreTokenType {
 }
 
 /// Python-exposed token subtype enum
-#[gen_stub_pyclass_enum]
-#[pyclass(name = "TokenSubType", module = "formualizer")]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pyclass_enum)]
+#[pyclass(
+    name = "TokenSubType",
+    module = "formualizer.formualizer_py",
+    from_py_object
+)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum PyTokenSubType {
     #[pyo3(name = "None_")]
@@ -109,7 +118,7 @@ impl Display for PyTokenSubType {
     }
 }
 
-#[gen_stub_pymethods]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pymethods)]
 #[pymethods]
 impl PyTokenSubType {
     fn __str__(&self) -> &'static str {
@@ -167,15 +176,19 @@ impl From<PyTokenSubType> for CoreTokenSubType {
 }
 
 /// Formula dialect enum exposed to Python
-#[gen_stub_pyclass_enum]
-#[pyclass(name = "FormulaDialect", module = "formualizer")]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pyclass_enum)]
+#[pyclass(
+    name = "FormulaDialect",
+    module = "formualizer.formualizer_py",
+    from_py_object
+)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum PyFormulaDialect {
     Excel,
     OpenFormula,
 }
 
-#[gen_stub_pymethods]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pymethods)]
 #[pymethods]
 impl PyFormulaDialect {
     fn __str__(&self) -> &'static str {
@@ -209,15 +222,19 @@ impl From<CoreFormulaDialect> for PyFormulaDialect {
 }
 
 /// Workbook mode enum exposed to Python.
-#[gen_stub_pyclass_enum]
-#[pyclass(name = "WorkbookMode", module = "formualizer")]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pyclass_enum)]
+#[pyclass(
+    name = "WorkbookMode",
+    module = "formualizer.formualizer_py",
+    from_py_object
+)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum PyWorkbookMode {
     Ephemeral,
     Interactive,
 }
 
-#[gen_stub_pymethods]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pymethods)]
 #[pymethods]
 impl PyWorkbookMode {
     fn __str__(&self) -> &'static str {
