+++ bindings/python/src/errors.rs
@@ -4,6 +4,7 @@ use formualizer::common::error::{
 use pyo3::exceptions::PyException;
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
+#[cfg(not(target_os = "emscripten"))]
 use pyo3_stub_gen::derive::{gen_stub_pyclass, gen_stub_pymethods};
 
 // Create custom exception types
@@ -126,14 +127,18 @@ impl ParserError {
 }
 
 /// Python representation of Excel domain errors
-#[gen_stub_pyclass]
-#[pyclass(name = "ExcelError", module = "formualizer")]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pyclass)]
+#[pyclass(
+    name = "ExcelError",
+    module = "formualizer.formualizer_py",
+    from_py_object
+)]
 #[derive(Clone, Debug)]
 pub struct PyExcelError {
     pub(crate) inner: RustExcelError,
 }
 
-#[gen_stub_pymethods]
+#[cfg_attr(not(target_os = "emscripten"), gen_stub_pymethods)]
 #[pymethods]
 impl PyExcelError {
     /// Create a new Excel error
